Update README.md
Absurdle is a twist on the popular game Wordle, designed to be as evil as possible. Rather than choosing a secret word at the start, Absurdle dynamically adapts to your guesses, always keeping the set of possible answers as large as possible. The goal is to prolong the game and resist being solved.